Gui WIP: Tab ordering

+++ zebra-crosslink/zebrad/src/commands/start.rs
@@ -618,6 +618,8 @@ impl StartCmd {
         let old_databases_task_handle_fused = (&mut old_databases_task_handle).fuse();
         pin!(old_databases_task_handle_fused);
 
+        if true
+        // if false // @Phillip @Testing
         {
             let tmp_dir = tempfile::TempDir::new().unwrap();
             let _ = std::fs::remove_dir_all(tmp_dir.path());
